Fixes backwards compatibility issue for depInj configration
* Fixes a backwards compatibility issue in the bundle configuration.php
file that let to problems with symfony versions in the 3.x range. The
problem manifested in not being able to read the configration and so
blocking the use of the symfony bin/console command. This patch
refactores the configuration code to check for the getRootNode()
function first and in a negative case it will prepare a default root
node value which is the bundles root node from the treeBuilder.